+++ src/main/java/racingcar/domain/Cars.java
@@ -1,4 +1,6 @@
-package racingcar.model;
+package racingcar.domain;
+
+import racingcar.domain.Car;
 
 import java.util.ArrayList;
 import java.util.List;
